// gen/proto/example/example_connect.ts
// @generated by protoc-gen-connect-es v0.13.0 with parameter "target=ts"
// @generated from file proto/example/example.proto (package example, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { CreateItemRequest, CreateItemResponse, DeleteItemRequest, GetItemRequest, GetItemResponse, ListItemsRequest, ListItemsResponse, UpdateItemRequest, UpdateItemResponse } from "./example_pb.js";
import { Empty, MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service example.ExampleService
 */
export const ExampleService = {
  typeName: "example.ExampleService",
  methods: {
    /**
     * @generated from rpc example.ExampleService.CreateItem
     */
    createItem: {
      name: "CreateItem",
      I: CreateItemRequest,
      O: CreateItemResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc example.ExampleService.GetItem
     */
    getItem: {
      name: "GetItem",
      I: GetItemRequest,
      O: GetItemResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc example.ExampleService.UpdateItem
     */
    updateItem: {
      name: "UpdateItem",
      I: UpdateItemRequest,
      O: UpdateItemResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc example.ExampleService.DeleteItem
     */
    deleteItem: {
      name: "DeleteItem",
      I: DeleteItemRequest,
      O: Empty,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc example.ExampleService.ListItems
     */
    listItems: {
      name: "ListItems",
      I: ListItemsRequest,
      O: ListItemsResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;